#54 - refactor: 키바나 관련 로직 컨트롤러, 서비스, 레포 통합
- ChartController.java, ElasticsearchRepository.java, TransactionAnalyzeService.java 제거
