+++ generator-api/src/main/java/com/simpaylog/generatorapi/service/TransactionExportService.java
@@ -2,9 +2,9 @@ package com.simpaylog.generatorapi.service;
 
 import com.simpaylog.generatorapi.exception.ApiException;
 import com.simpaylog.generatorapi.exception.ErrorCode;
+import com.simpaylog.generatorapi.repository.Elasticsearch.TransactionAggregationRepository;
 import com.simpaylog.generatorcore.enums.export.ExportFormat;
 import com.simpaylog.generatorcore.exception.CoreException;
-import com.simpaylog.generatorapi.repository.Elasticsearch.ElasticsearchRepository;
 import com.simpaylog.generatorapi.utils.FileExporter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +20,7 @@ import static com.simpaylog.generatorapi.exception.ErrorCode.INVALID_EXPORT_FORM
 @RequiredArgsConstructor
 public class TransactionExportService {
 
-    private final ElasticsearchRepository repository;
+    private final TransactionAggregationRepository repository;
     private final FileExporter fileExporter;
 
     // TODO: sessionId 없는 경우 예외처리 필요
